# Python script to compare commits with llama-bench