This PR adds a polished version of a Python script that I have used privately to create tables that compare the performance of two commits by writing them to SQLite as an intermediary. This is an example table:

| Model         | GPU                     |   Model Size [GiB] | Test   |   t/s master |   t/s cuda-mmqf-10 |   Speedup |
|---------------|-------------------------|--------------------|--------|--------------|--------------------|-----------|
| llama 7B Q4_0 | NVIDIA GeForce RTX 3090 |               3.56 | pp512  |      3890.69 |            2236.42 |      0.57 |
| llama 7B Q4_0 | NVIDIA GeForce RTX 3090 |               3.56 | tg128  |       137.05 |             136.58 |      1.00 |
| llama 7B Q8_0 | NVIDIA GeForce RTX 3090 |               6.67 | pp512  |      3809.83 |            3026.61 |      0.79 |
| llama 7B Q8_0 | NVIDIA GeForce RTX 3090 |               6.67 | tg128  |        90.09 |              89.66 |      1.00 |

To avoid repeating myself, I'm simply copy-pasting the help:

```
usage: compare-llama-bench.py [-h] [-b BASELINE] [-c COMPARE] [-i INPUT] [-o OUTPUT] [-s SHOW]

Creates tables from llama-bench data written to an SQLite database. Example usage (Linux):

$ git checkout master
$ make clean && make llama-bench
$ ./llama-bench -o sql | sqlite3 llama-bench.sqlite
$ git checkout some_branch
$ make clean && make llama-bench
$ ./llama-bench -o sql | sqlite3 llama-bench.sqlite
$ ./scripts/compare-llama-bench.py

Performance numbers from multiple runs per commit are averaged WITHOUT being weighted by the --repetitions parameter of llama-bench.

options:
  -h, --help            show this help message and exit
  -b BASELINE, --baseline BASELINE
                        The baseline commit to compare performance to. Accepts either a branch name, tag name, or commit hash. Defaults to latest
                        master commit with data.
  -c COMPARE, --compare COMPARE
                        The commit whose performance is to be compared to the baseline. Accepts either a branch name, tag name, or commit hash.
                        Defaults to the non-master commit for which llama-bench was run most recently.
  -i INPUT, --input INPUT
                        Input SQLite file for comparing commits. Defaults to 'llama-bench.sqlite' in the current working directory. If no such
                        file is found and there is exactly one .sqlite file in the current directory, that file is instead used as input.
  -o OUTPUT, --output OUTPUT
                        Output format for the table. Defaults to 'github'. Also supports e.g. 'latex' or 'mediawiki'. See tabulate documentation
                        for full list.
  -s SHOW, --show SHOW  Columns to add to the table. Accepts a comma-separated list of values. Legal values: cuda, opencl, metal, gpu_blas, blas,
                        cpu_info, gpu_info, model_filename, model_type, model_size, model_n_params, n_batch, n_threads, type_k, type_v,
                        n_gpu_layers, main_gpu, no_kv_offload, mul_mat_q, tensor_split. Defaults to model name (model_type) and CPU and/or GPU
                        name (cpu_info, gpu_info) plus any column where not all data points are the same. If the columns are manually specified,
                        then the results for each unique combination of the specified values are averaged WITHOUT weighing by the --repetitions
                        parameter of llama-bench.
```