@JohannesGaessler 2024-01-09
I chose not to display the file names by default because those aren't useful information for making tables but looking at it again model size is also kind of redundant because you can infer it from the model name. So maybe that one should be hidden by default too?