@slaren 2024-01-09
I don't have a strong opinion about that, the model size can still be useful to remind us, but it is redudant. It would be nice if the columns were a bit briefer, on my screen this shows two lines per row because it doesn't fit all the text. This fits:

| Model         | GPU                        |  Size [GiB] | Test   |   t/s master |   t/s sl/backend-sched |   Speedup |
|---------------|----------------------------|--------------------|--------|--------------|------------------------|-----------|
| llama 7B F16  | RTX 3090 Ti |              12.55 | pp512  |      4875.12 |                5424.32 |      1.11 |
| llama 7B F16  | RTX 3090 Ti |              12.55 | tg128  |        55.19 |                  50.47 |      0.91 |
| llama 7B Q4_0 | RTX 3090 Ti |               3.56 | pp512  |      4048.87 |                4368.36 |      1.08 |
| llama 7B Q4_0 | RTX 3090 Ti |               3.56 | tg128  |       126.22 |                 125.61 |      1.00 |