@JohannesGaessler 2024-01-09
I just implemented some simple logic to strip the prefixes from GPU names; that should make the tables small enough.