@slaren 2024-01-09
`-o pipe` prints the column alignment colons, and also works on github, so it may be a better default.

| GPU         | Model         |   Model Size [GiB] | Test   |   t/s master |   t/s sl/backend-sched |   Speedup |
|:------------|:--------------|-------------------:|:-------|-------------:|-----------------------:|----------:|
| RTX 3090 Ti | llama 7B F16  |              12.55 | pp512  |      4875.12 |                5394.60 |      1.11 |
| RTX 3090 Ti | llama 7B F16  |              12.55 | tg128  |        55.19 |                  50.37 |      0.91 |
| RTX 3090 Ti | llama 7B Q4_0 |               3.56 | pp512  |      4048.87 |                4369.56 |      1.08 |
| RTX 3090 Ti | llama 7B Q4_0 |               3.56 | tg128  |       126.22 |                 125.72 |      1.00 |